+++ spec/modl/modl2.rb
@@ -41,14 +41,22 @@ end
 
 a = <<~XXXXXXXX
 
-  _test1=e1afmkfd;
-  russian1=%test1.p;
-
-  _test2="e1afmkfd";
-  russian2=%test2.p;
-
-  _test3=`e1afmkfd`;
-  russian3=%test3.p;
+  _co=at;
+  _l=de;
+  {
+   co=at?
+     country=Austria
+     language={
+       l=fr?
+         French
+       /l=de?
+         German
+       /?
+         Other
+     }
+   /?
+     country=Other
+  }
 
 XXXXXXXX