+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Use file cache on load error - GitHub issue #20
 - Error converting value to number - GitHub issue #21
 - Problem with punycode - GitHub issue #22
+- Handle orphan pairs - adopt then into a map
 
 0.3.11
 ===